Fix nits / rename file
