+++ python/ray/serve/tests/test_serve_ha.py
@@ -109,7 +109,7 @@ def test_ray_serve_basic(docker_cluster):
 import ray
 import requests
 from ray.serve.schema import ServeInstanceDetails
-from ray._private.resource_spec import HEAD_NODE_RESOURCE_NAME
+from ray._private.resource_and_label_spec import HEAD_NODE_RESOURCE_NAME
 ray.init(address="auto")
 head_node_id = ray.get_runtime_context().get_node_id()
 serve_details = ServeInstanceDetails(
